+++ app/Http/Controllers/Company/CompanyDeleteController.php
@@ -7,12 +7,16 @@ namespace App\Http\Controllers\Company;
 use App\Http\Controllers\MainController;
 use App\Models\Company;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CompanyDeleteController extends MainController
 {
     public function delete(Request $request, int $id)
     {
         $company = Company::find($id);
+        if (Auth::user()->company_id !== 1) {
+            return response()->json(['message' => 'Unauthorized'], 401);
+        }
         $company->delete();
         $company->save();
